tmp: array_api dtypes

+++ scipy/_lib/_array_api.py
@@ -271,7 +271,7 @@ def xp_assert_equal(actual, desired, *, check_namespace=True, check_dtype=True,
     if xp is None:
         xp = array_namespace(actual)
     if isinstance(desired, list):
-        desired = xp.asarray(desired)
+        desired = xp.asarray(desired, dtype=actual.dtype)
 
     actual, desired = _strict_check(
         actual, desired, xp, check_namespace=check_namespace,
@@ -298,7 +298,7 @@ def xp_assert_close(actual, desired, *, rtol=None, atol=0, check_namespace=True,
     if xp is None:
         xp = array_namespace(actual)
     if isinstance(desired, list):
-        desired = xp.asarray(desired)
+        desired = xp.asarray(desired, dtype=actual.dtype)
 
     actual, desired = _strict_check(
         actual, desired, xp,
